Moved exceptions back to where they're thrown
