--- javaspec-runner/src/main/java/info/javaspec/spec/OuterClassFactory.java
@@ -1,15 +1,0 @@
-package info.javaspec.spec;
-
-import java.lang.reflect.Constructor;
-
-final class OuterClassFactory extends ClassFactory {
-  @Override
-  protected Constructor<?> getConstructor(Class<?> outerClass) throws NoSuchMethodException {
-    return outerClass.getDeclaredConstructor();
-  }
-
-  @Override
-  protected Object makeInstance(Constructor<?> constructor) throws ReflectiveOperationException {
-    return constructor.newInstance();
-  }
-}
