+++ javaspec-runner/src/main/java/info/javaspec/spec/SpecExecutionContext.java
@@ -1,5 +1,6 @@
 package info.javaspec.spec;
 
+import java.lang.reflect.Constructor;
 import java.lang.reflect.Field;
 import java.lang.reflect.Modifier;
 import java.util.HashMap;
@@ -46,6 +47,52 @@ final class SpecExecutionContext {
     } else {
       Object enclosingObject = makeAndRememberInstance(enclosingClass);
       return new InnerClassFactory(enclosingClass, enclosingObject).makeInstance(contextClass);
+    }
+  }
+
+  static final class OuterClassFactory extends ClassFactory {
+    @Override
+    protected Constructor<?> getConstructor(Class<?> outerClass) throws NoSuchMethodException {
+      return outerClass.getDeclaredConstructor();
+    }
+
+    @Override
+    protected Object makeInstance(Constructor<?> constructor) throws ReflectiveOperationException {
+      return constructor.newInstance();
+    }
+  }
+
+  static final class InnerClassFactory extends ClassFactory {
+    private final Class<?> enclosingClass;
+    private final Object enclosingObject;
+
+    public InnerClassFactory(Class<?> enclosingClass, Object enclosingObject) {
+      this.enclosingClass = enclosingClass;
+      this.enclosingObject = enclosingObject;
+    }
+
+    @Override
+    protected Constructor<?> getConstructor(Class<?> innerClass) throws NoSuchMethodException {
+      return innerClass.getDeclaredConstructor(enclosingClass);
+    }
+
+    @Override
+    protected Object makeInstance(Constructor<?> constructor) throws ReflectiveOperationException {
+      return constructor.newInstance(enclosingObject);
+    }
+  }
+
+  static final class TestSetupFailed extends RuntimeException {
+    private static final long serialVersionUID = 1L;
+
+    public static TestSetupFailed forClass(Class<?> context, Throwable cause) {
+      return new TestSetupFailed(
+        String.format("Failed to create test context %s", context.getName()),
+        cause);
+    }
+
+    private TestSetupFailed(String message, Throwable cause) {
+      super(message, cause);
     }
   }
 }
