+++ javaspec-runner/src/test/java/info/javaspec/spec/SpecFactoryTest.java
@@ -1,10 +1,10 @@
 package info.javaspec.spec;
 
 import de.bechte.junit.runners.context.HierarchicalContextRunner;
-import info.javaspec.context.AmbiguousFixture;
 import info.javaspec.context.Context;
 import info.javaspec.context.ContextFactory;
 import info.javaspec.context.FakeContext;
+import info.javaspec.spec.SpecFactory.AmbiguousFixture;
 import info.javaspecproto.ContextClasses;
 import org.junit.After;
 import org.junit.Before;
